rename files to match angular convention
