+++ frontend/src/app/interface/card-data.ts
