+++ frontend/src/app/interface/stats-response.ts
